category page all section spliting by component
