+++ src/app/admin/category/_component/DeleteCategoryModal.tsx
@@ -1,0 +1,42 @@
+import { Button } from "@/components/ui/button";
+import {
+  Dialog,
+  DialogContent,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+} from "@/components/ui/dialog";
+
+type TProps = {
+  isDeleteModalOpen: boolean;
+  setIsDeleteModalOpen: React.Dispatch<React.SetStateAction<boolean>>;
+};
+
+export default function DeleteCategoryModal({
+  isDeleteModalOpen,
+  setIsDeleteModalOpen,
+}: TProps) {
+  const handleDeleteCategory = () => {};
+  return (
+    <Dialog open={isDeleteModalOpen} onOpenChange={setIsDeleteModalOpen}>
+      <DialogContent className="sm:max-w-[425px]">
+        <DialogHeader>
+          <DialogTitle className="text-gray-900 dark:text-gray-100">
+            Delete Category
+          </DialogTitle>
+        </DialogHeader>
+        <p className="text-gray-900 dark:text-gray-100">
+          Are you sure you want to delete this category?
+        </p>
+        <DialogFooter>
+          <Button variant="outline" onClick={() => setIsDeleteModalOpen(false)}>
+            Cancel
+          </Button>
+          <Button variant="destructive" onClick={handleDeleteCategory}>
+            Delete
+          </Button>
+        </DialogFooter>
+      </DialogContent>
+    </Dialog>
+  );
+}
